fixed vision rectangles & started paths
