--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/bluevision/EmptyProcessorBlue.java
@@ -1,24 +1,0 @@
-package org.firstinspires.ftc.teamcode.bluevision;
-
-import android.graphics.Canvas;
-import org.firstinspires.ftc.robotcore.internal.camera.calibration.CameraCalibration;
-import org.firstinspires.ftc.vision.VisionProcessor;
-import org.opencv.core.Mat;
-
-public class EmptyProcessorBlue implements VisionProcessor {
-    @Override
-    public void init(int width, int height, CameraCalibration calibration) {
-
-    }
-
-    @Override
-    public Object processFrame(Mat frame, long captureTimeNanos) {
-        return null;
-    }
-
-    @Override
-    public void onDrawFrame(Canvas canvas, int onscreenWidth, int onscreenHeight, float scaleBmpPxToCanvasPx, float scaleCanvasDensity, Object userContext)
-    {
-    }
-
-}
