--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/bluevision/FirstVisionProcessorBlue.java
@@ -1,107 +1,0 @@
-package org.firstinspires.ftc.teamcode.bluevision;
-
-import android.graphics.Canvas;
-import android.graphics.Color;
-import android.graphics.Paint;
-
-import org.firstinspires.ftc.robotcore.internal.camera.calibration.CameraCalibration;
-import org.firstinspires.ftc.teamcode.redvision.FirstVisionProcessor;
-import org.firstinspires.ftc.vision.VisionProcessor;
-import org.opencv.core.Core;
-import org.opencv.core.Mat;
-import org.opencv.core.Rect;
-import org.opencv.core.Scalar;
-import org.opencv.imgproc.Imgproc;
-
-public class FirstVisionProcessorBlue implements VisionProcessor{
-    public Rect rectLeft = new Rect(0, 40, 175, 400);
-    public Rect rectMiddle = new Rect(200, 40, 250, 400);
-    public Rect rectRight = new Rect(450, 40, 175, 400);
-    FirstVisionProcessor.Selected selection = FirstVisionProcessor.Selected.NONE;
-
-    Mat submat = new Mat();
-    Mat hsvMat = new Mat();
-    @Override
-    public void init(int width, int height, CameraCalibration calibration) {
-    }
-
-    @Override
-    public Object processFrame(Mat frame, long captureTimeNanos) {
-        Imgproc.cvtColor(frame, hsvMat, Imgproc.COLOR_RGB2HSV);
-        double satRectLeft = getAvgSaturation(hsvMat, rectLeft);
-        double satRectMiddle = getAvgSaturation(hsvMat, rectMiddle);
-        double satRectRight = getAvgSaturation(hsvMat, rectRight);
-
-        if ((satRectLeft > satRectMiddle) && (satRectLeft > satRectRight)) {
-            return FirstVisionProcessor.Selected.LEFT;
-        } else if ((satRectMiddle > satRectLeft) && (satRectMiddle > satRectRight)) {
-            return FirstVisionProcessor.Selected.MIDDLE;
-        }
-        return FirstVisionProcessor.Selected.RIGHT;
-    }
-
-    protected double getAvgSaturation(Mat input, Rect rect) {
-        submat = input.submat(rect);
-        Scalar color = Core.mean(submat);
-        return color.val[1];
-    }
-
-    private android.graphics.Rect makeGraphicsRect(Rect rect, float scaleBmpPxToCanvasPx) {
-
-        int left = Math.round(rect.x * scaleBmpPxToCanvasPx);
-        int top = Math.round(rect.y * scaleBmpPxToCanvasPx);
-        int right = left + Math.round(rect.width * scaleBmpPxToCanvasPx);
-        int bottom = top + Math.round(rect.height * scaleBmpPxToCanvasPx);
-        return new android.graphics.Rect(left, top, right, bottom);
-    }
-
-    @Override
-    public void onDrawFrame(Canvas canvas, int onscreenWidth, int onscreenHeight,float scaleBmpPxToCanvasPx, float scaleCanvasDensity, Object userContext) {
-
-
-        Paint selectedPaint = new Paint();
-        selectedPaint.setColor(Color.RED);
-        selectedPaint.setStyle(Paint.Style.STROKE);
-        selectedPaint.setStrokeWidth(scaleCanvasDensity * 4);
-        Paint nonSelectedPaint = new Paint(selectedPaint);
-        nonSelectedPaint.setColor(Color.GREEN);
-
-        android.graphics.Rect drawRectangleLeft = makeGraphicsRect(rectLeft, scaleBmpPxToCanvasPx);
-        android.graphics.Rect drawRectangleMiddle = makeGraphicsRect(rectMiddle, scaleBmpPxToCanvasPx);
-        android.graphics.Rect drawRectangleRight = makeGraphicsRect(rectRight, scaleBmpPxToCanvasPx);
-
-        selection = (FirstVisionProcessor.Selected) userContext;
-        switch (selection) {
-            case LEFT:
-                canvas.drawRect(drawRectangleLeft, selectedPaint);
-                canvas.drawRect(drawRectangleMiddle, nonSelectedPaint);
-                canvas.drawRect(drawRectangleRight, nonSelectedPaint);
-                break;
-            case MIDDLE:
-                canvas.drawRect(drawRectangleLeft, nonSelectedPaint);
-                canvas.drawRect(drawRectangleMiddle, selectedPaint);
-                canvas.drawRect(drawRectangleRight, nonSelectedPaint);
-                break;
-            case RIGHT:
-                canvas.drawRect(drawRectangleLeft, nonSelectedPaint);
-                canvas.drawRect(drawRectangleMiddle, nonSelectedPaint);
-                canvas.drawRect(drawRectangleRight, selectedPaint);
-                break;
-            case NONE:
-                canvas.drawRect(drawRectangleLeft, nonSelectedPaint);
-                canvas.drawRect(drawRectangleMiddle, nonSelectedPaint);
-                canvas.drawRect(drawRectangleRight, nonSelectedPaint);
-                break;
-        }
-    }
-    public FirstVisionProcessor.Selected getSelection() {
-        return selection;
-    }
-
-    public enum Selected {
-        NONE,
-        LEFT,
-        MIDDLE,
-        RIGHT
-    }
-}
