--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/redvision/ThreeRectanglesProcessor.java
@@ -1,81 +1,0 @@
-package org.firstinspires.ftc.teamcode.redvision;
-
-import android.graphics.Canvas;
-import android.graphics.Color;
-import android.graphics.Paint;
-
-import org.firstinspires.ftc.robotcore.internal.camera.calibration.CameraCalibration;
-import org.firstinspires.ftc.vision.VisionProcessor;
-import org.opencv.core.Mat;
-import org.opencv.core.Rect;
-
-public class ThreeRectanglesProcessor implements VisionProcessor {
-public Rect rectLeft = new Rect(110, 60, 300, 240);
-public Rect rectMiddle = new Rect(160, 60, 300, 240);
-public Rect rectRight = new Rect(210, 60, 300, 240);
-Selected selection = Selected.NONE;
-
-    @Override
-    public void init(int width, int height, CameraCalibration calibration) {
-         }
-
-        @Override
-        public Object processFrame(Mat frame, long captureTimeNanos) {
-            return null;
-         }
-
-        private android.graphics.Rect makeGraphicsRect(Rect rect, float scaleBmpPxToCanvasPx) {
-        int left = Math.round(rect.x * scaleBmpPxToCanvasPx);
-        int top = Math.round(rect.y * scaleBmpPxToCanvasPx);
-        int right = left + Math.round(rect.width * scaleBmpPxToCanvasPx);
-        int bottom = top + Math.round(rect.height * scaleBmpPxToCanvasPx);
-
-            return new android.graphics.Rect(left, top, right, bottom);
-        }
-
-        @Override
-        public void onDrawFrame(Canvas canvas, int onscreenWidth, int onscreenHeight, float scaleBmpPxToCanvasPx, float scaleCanvasDensity, Object userContext) {
-
-        Paint selectedPaint = new Paint();
-        selectedPaint.setColor(Color.RED);
-        selectedPaint.setStyle(Paint.Style.STROKE);
-        selectedPaint.setStrokeWidth(scaleCanvasDensity * 4);
-
-        Paint nonSelectedPaint = new Paint(selectedPaint);
-        nonSelectedPaint.setColor(Color.GREEN);
-
-        android.graphics.Rect drawRectangleLeft = makeGraphicsRect(rectLeft, scaleBmpPxToCanvasPx);
-        android.graphics.Rect drawRectangleMiddle = makeGraphicsRect(rectMiddle, scaleBmpPxToCanvasPx);
-        android.graphics.Rect drawRectangleRight = makeGraphicsRect(rectRight, scaleBmpPxToCanvasPx);
-
-        switch (selection) {
-            case LEFT:
-                canvas.drawRect(drawRectangleLeft, selectedPaint);
-                canvas.drawRect(drawRectangleMiddle, nonSelectedPaint);
-                canvas.drawRect(drawRectangleRight, nonSelectedPaint);
-                break;
-            case MIDDLE:
-                canvas.drawRect(drawRectangleLeft, nonSelectedPaint);
-                canvas.drawRect(drawRectangleMiddle, selectedPaint);
-                canvas.drawRect(drawRectangleRight, nonSelectedPaint);
-                break;
-            case RIGHT:
-                canvas.drawRect(drawRectangleLeft, nonSelectedPaint);
-                canvas.drawRect(drawRectangleMiddle, nonSelectedPaint);
-                canvas.drawRect(drawRectangleRight, selectedPaint);
-                break;
-            case NONE:
-                canvas.drawRect(drawRectangleLeft, nonSelectedPaint);
-                canvas.drawRect(drawRectangleMiddle, nonSelectedPaint);
-                canvas.drawRect(drawRectangleRight, nonSelectedPaint);
-                break;
-             }
-         }
-
-        public enum Selected {
-                NONE,
-                LEFT,
-                MIDDLE,
-                RIGHT
-    }
-}
