+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/vision/EmptyProcessor.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.redvision;
+package org.firstinspires.ftc.teamcode.vision;
 
 import android.graphics.Canvas;
 import org.firstinspires.ftc.robotcore.internal.camera.calibration.CameraCalibration;
